% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2Plot.R
\name{AddMeanCrossBar}
\alias{AddMeanCrossBar}
\title{AddMeanCrossBar}
\usage{
AddMeanCrossBar(fun = mean, width = 0.5, position = "identity", ...)
}
\arguments{
\item{fun}{mean}

\item{width}{0.5}

\item{position}{identity}

\item{...}{}
}
\value{

}
\description{
AddMeanCrossBar
}
