% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportPDF.R
\name{ExportPDF_TiffTolocal_withData}
\alias{ExportPDF_TiffTolocal_withData}
\title{ExportPDF_TiffTolocal_withData}
\usage{
ExportPDF_TiffTolocal_withData(
  p,
  filename = "Expot",
  width_screen = 500,
  Heigh_screen = 500,
  pig_device = "tiff",
  units = c("in", "cm", "mm"),
  dpi = 92,
  col_names = TRUE,
  row_names = FALSE,
  format_headers = TRUE,
  use_zip64 = FALSE
)
}
\arguments{
\item{use_zip64}{}
}
\value{

}
\description{
ExportPDF_TiffTolocal_withData
}
