% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2Plot.R
\name{reorderXtiticks}
\alias{reorderXtiticks}
\title{reorderXtiticks
ggplot2排序X轴标签}
\usage{
reorderXtiticks(fun = mean, decreasing = FALSE, ...)
}
\arguments{
\item{fun}{mean}

\item{decreasing}{FALSE}

\item{...}{}
}
\value{

}
\description{
reorderXtiticks
ggplot2排序X轴标签
}
\examples{
ggplot(pdata,aes(RareQcluster\%>\%as.character(),PTPRC))+geom_violin()+reorderXtiticks(fun = mean,decreasing = F)
}
