% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrublet.R
\name{run_scrublet}
\alias{run_scrublet}
\title{run_scrublet}
\usage{
run_scrublet(counts_matrix, env_name = "r-reticulate")
}
\arguments{
\item{counts_matrix}{genes x cells}

\item{env_name}{}
}
\value{

}
\description{
run_scrublet
}
