% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrublet.R
\name{run_scrublet_seurat}
\alias{run_scrublet_seurat}
\title{run_scrublet_seurat}
\usage{
run_scrublet_seurat(
  seurat_obj,
  split.by = NULL,
  assay = "RNA",
  env_name = "r-reticulate"
)
}
\arguments{
\item{env_name}{}
}
\value{

}
\description{
run_scrublet_seurat
}
